Added precipitation_type subpackage and moved appropriate modules.
