+++ improver/cli/convection_ratio.py
@@ -54,7 +54,7 @@ def process(*cubes: cli.inputcube,):
             A cube of convection_ratio of the same dimensions as the input cubes.
 
     """
-    from improver.convection import ConvectionRatioFromComponents
+    from improver.precipitation_type.convection import ConvectionRatioFromComponents
 
     if len(cubes) != 2:
         raise IOError(f"Expected 2 input cubes, received {len(cubes)}")
