+++ improver/cli/sleet_probability.py
@@ -53,7 +53,9 @@ def process(snow: cli.inputcube, rain: cli.inputcube):
             Returns a cube with the probability of sleet.
     """
 
-    from improver.calculate_sleet_prob import calculate_sleet_probability
+    from improver.precipitation_type.calculate_sleet_prob import (
+        calculate_sleet_probability
+    )
 
     result = calculate_sleet_probability(snow, rain)
     return result
