+++ improver/precipitation_type/snow_fraction.py
