+++ improver_tests/calculate_sleet_prob/test_calculate_sleet_probability.py
@@ -35,7 +35,9 @@ import unittest
 import numpy as np
 from iris.tests import IrisTest
 
-from improver.calculate_sleet_prob import calculate_sleet_probability
+from improver.precipitation_type.calculate_sleet_prob import (
+    calculate_sleet_probability
+)
 from improver.synthetic_data.set_up_test_cubes import set_up_probability_cube
