+++ improver_tests/convection/test_ConvectionRatioFromComponents.py
@@ -36,7 +36,7 @@ import pytest
 from iris.cube import CubeList
 from numpy.testing import assert_allclose, assert_equal, assert_raises_regex
 
-from improver.convection import ConvectionRatioFromComponents
+from improver.precipitation_type.convection import ConvectionRatioFromComponents
 from improver.synthetic_data.set_up_test_cubes import set_up_variable_cube
 
 GLOBAL_ATTRIBUTES = {
